pyproject.toml
[project]
name = "git-utils"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython==3.1.46",
    "pydbus>=0.6.0",
    "pygithub==2.8.1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "rich==14.2.0",
    "rich-argparse==1.7.2",
    "xdg==6.0.0",
    "xdg-base-dirs>=6.0.2",
]

[project.scripts]
git-link = "gitutils.link:main"
git-release-branch = "gitutils.release:main"

[tool.uv]
package = true
